+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""DohaAudio test support package."""